+++ webknossos/tests/dataset/test_remote_dataset.py
@@ -10,6 +10,8 @@ from upath import UPath
 from tests.constants import TESTOUTPUT_DIR
 from webknossos import (
     COLOR_CATEGORY,
+    AgglomerateAttachment,
+    AgglomerateGraph,
     BoundingBox,
     DataFormat,
     Dataset,
@@ -20,6 +22,7 @@ from webknossos import (
     RemoteFolder,
     Team,
     TransferMode,
+    Vec3Int,
 )
 from webknossos.utils import is_remote_path, rmtree
 
@@ -35,7 +38,16 @@ SAMPLE_BBOX = BoundingBox((3164, 3212, 1017), (10, 10, 10))
 def sample_downloaded_dataset(tmp_upath: UPath) -> Iterator[Dataset]:
     yield RemoteDataset.open("l4_sample").download(
         path=tmp_upath / "l4_sample",
-        bbox=BoundingBox((3457, 3323, 1204), (10, 10, 10)),
+        bounding_box=BoundingBox((3457, 3323, 1204), (10, 10, 10)),
+    )
+
+
+def get_sample_dataset(
+    tmpdir: UPath, *, layers: list[str] | None = None, bbox: BoundingBox = SAMPLE_BBOX
+) -> Dataset:
+    url = "http://localhost:9000/datasets/Organization_X/l4_sample"
+    return RemoteDataset.open(url).download(
+        path=UPath(tmpdir) / "sample_ds", bounding_box=bbox, layers=layers
     )
 
 
@@ -53,7 +65,7 @@ def _prepare_dataset_path(output_path: UPath, suffix: str) -> UPath:
 
 
 @pytest.mark.skip(
-    reason="This could work in CI with a local minio instance. Configuring webknossos is a bit more involved and, therefore, future work."
+    reason="TransferMode.COPY requires absolute paths, which are different on multiple machines. Skipping, for now."
 )
 def test_remote_dataset_add_layer_as_copy() -> None:
     ds_path = _prepare_dataset_path(TESTOUTPUT_DIR, "remote_copy_src")
@@ -83,7 +95,7 @@ def test_remote_dataset_add_layer_as_copy() -> None:
 
 
 @pytest.mark.skip(
-    reason="This could work in CI with a local minio instance. Configuring webknossos is a bit more involved and, therefore, future work."
+    reason="TransferMode.COPY requires absolute paths, which are different on multiple machines. Skipping, for now."
 )
 def test_remote_dataset_add_mag_as_copy() -> None:
     ds_path = _prepare_dataset_path(TESTOUTPUT_DIR, "remote_copy_mag_src")
@@ -214,33 +226,6 @@ def test_shallow_copy_remote_layers(tmp_upath: UPath) -> None:
     assert data.shape == (1, 64, 64, 256)
 
 
-def test_add_mag_ref_from_local_path(tmp_upath: UPath) -> None:
-    dataset1 = Dataset(tmp_upath / "origin", voxel_size=(10, 10, 10))
-    dataset1.write_layer(
-        "color",
-        COLOR_CATEGORY,
-        data=np.ones((1, 10, 10, 10), dtype="uint8"),
-        downsample=False,
-    )
-
-    dataset2 = Dataset(tmp_upath / "copy", voxel_size=(10, 10, 10))
-    layer1 = dataset2.add_layer_as_ref(tmp_upath / "origin" / "color")
-    layer1_mag1 = layer1.get_mag(1)
-
-    assert layer1_mag1.path == tmp_upath / "origin" / "color" / "1"
-    assert layer1_mag1._properties.path == str(
-        (tmp_upath / "origin" / "color" / "1").resolve()
-    )
-
-    layer2_mag1 = dataset2.add_layer("color2", COLOR_CATEGORY).add_mag_as_ref(
-        tmp_upath / "origin" / "color" / "1"
-    )
-    assert layer2_mag1.path == tmp_upath / "origin" / "color" / "1"
-    assert layer2_mag1._properties.path == str(
-        (tmp_upath / "origin" / "color" / "1").resolve()
-    )
-
-
 def test_changing_properties_on_remote_dataset() -> None:
     remote_dataset = RemoteDataset.open(dataset_id="59e9cfbdba632ac2ab8b23b5")
     remote_dataset.description = "This is a test description"
@@ -277,13 +262,6 @@ def test_changing_properties_on_read_only_remote_dataset() -> None:
     assert remote_dataset.description == description_before_change_attempt
 
 
-def get_sample_dataset(tmpdir: UPath) -> Dataset:
-    url = "http://localhost:9000/datasets/Organization_X/l4_sample"
-    return RemoteDataset.open(url).download(
-        path=UPath(tmpdir) / "sample_ds", bbox=SAMPLE_BBOX
-    )
-
-
 def test_get_remote_datasets() -> None:
     datasets = RemoteDataset.list()
     assert any(ds.name == "l4_sample" for ds in datasets.values())
@@ -308,14 +286,12 @@ def test_get_remote_datasets() -> None:
     [
         "http://localhost:9000/datasets/Organization_X/l4_sample",
         "http://localhost:9000/datasets/Organization_X/l4_sample/view",
-        # "http://localhost:9000/datasets/scalable_minds/l4_sample_dev_sharing/view?token=ilDXmfQa2G8e719vb1U9YQ#%7B%22orthogonal%7D",
-        # "http://localhost:9000/links/93zLg9U9vJ3c_UWp",
     ],
 )
 def test_url_download(url: str, tmp_upath: UPath) -> None:
     sample_dataset = get_sample_dataset(tmp_upath)
     ds = RemoteDataset.open(url).download(
-        path=tmp_upath / "ds", mags=[Mag(1)], bbox=SAMPLE_BBOX
+        path=tmp_upath / "ds", mags=[Mag(1)], bounding_box=SAMPLE_BBOX
     )
     assert set(ds.layers.keys()) == {"color", "segmentation"}
     data = ds.get_color_layers()[0].get_finest_mag().read()
@@ -361,7 +337,7 @@ def test_url_open_remote(
 
 
 @pytest.mark.skip(
-    reason="This won't work in CI as the paths stored in cassettes are always absolute and dependent on the system recording the cassette."
+    reason="TransferMode.MOVE_AND_SYMLINK requires absolute paths, which are different on multiple machines. Skipping, for now."
 )
 def test_upload_dataset_with_symlinks(tmp_upath: UPath) -> None:
     sample_dataset = get_sample_dataset(tmp_upath)
@@ -376,7 +352,7 @@ def test_upload_dataset_with_symlinks(tmp_upath: UPath) -> None:
 
 
 @pytest.mark.skip(
-    reason="This could work in CI with a local minio instance. Configuring webknossos is a bit more involved and, therefore, future work."
+    reason="TransferMode.COPY requires absolute paths, which are different on multiple machines. Skipping, for now."
 )
 def test_upload_dataset_copy_to_paths(tmp_upath: UPath) -> None:
     sample_dataset = get_sample_dataset(tmp_upath)
@@ -462,10 +438,6 @@ def test_upload_download_roundtrip(tmp_upath: UPath) -> None:
     uploaded_dataset = ds_original.upload(
         new_dataset_name="test_upload_download_roundtrip"
     )
-    RemoteDataset.trigger_reload_in_datastore(
-        dataset_name_or_url="test_upload_download_roundtrip",
-        organization_id="Organization_X",
-    )
     ds_roundtrip = RemoteDataset.open(uploaded_dataset.url).download(
         path=tmp_upath / "ds", layers=["color", "segmentation"]
     )
@@ -484,3 +456,39 @@ def test_upload_twice(tmp_upath: UPath) -> None:
     remote2 = ds_original.upload(new_dataset_name="test_upload_twice")
     assert remote1.url != remote2.url
     assert remote1.name == remote2.name
+
+
+@pytest.mark.skip(
+    reason="This could work in CI with a local minio instance. Configuring webknossos is a bit more involved and, therefore, future work."
+)
+def test_add_attachment(tmp_upath: UPath) -> None:
+    ds_original = get_sample_dataset(
+        tmp_upath,
+        layers=["segmentation"],
+        bbox=SAMPLE_BBOX.with_size(Vec3Int(32, 32, 32)),
+    )
+
+    seg_layer = ds_original.get_layer("segmentation").as_segmentation_layer()
+    seg_data = seg_layer.get_finest_mag().read()
+    seg_ids = np.unique(seg_data)
+
+    # Remap segment ids to be dense
+    remapped_seg_ids = {seg_id: i + 1 for i, seg_id in enumerate(seg_ids)}
+    for old_id, new_id in remapped_seg_ids.items():
+        seg_data[seg_data == old_id] = new_id
+    seg_layer.get_finest_mag().write(seg_data)
+
+    # Construct agglomerate graph and attachment
+    seg_ids = np.unique(seg_data)
+    graph = AgglomerateGraph()
+    for seg_id in seg_ids:
+        seg_position = seg_layer.bounding_box.topleft_xyz + Vec3Int(
+            np.argwhere(seg_data == seg_id)[0][1:]
+        )
+        graph.add_segment(int(seg_id), position=seg_position)
+    for seg_id in seg_ids[1:]:
+        graph.add_affinity_edge(int(seg_id), int(seg_ids[0]), affinity=0.5)
+    AgglomerateAttachment.create_and_add_to(seg_layer, "map_all", graph)
+
+    # Upload dataset
+    ds_original.upload(new_dataset_name="test_agglomerate_attachment")