+++ webknossos/webknossos/cli/upload.py
@@ -72,9 +72,10 @@ def main(
         folder_obj: None | RemoteFolder = None
         if folder is not None:
             folder_obj = RemoteFolder.get_by_path(folder)
-        Dataset.open(dataset_path=source).upload(
+        uploaded_dataset = Dataset.open(dataset_path=source).upload(
             new_dataset_name=dataset_name,
             jobs=jobs,
             folder=folder_obj,
             transfer_mode=transfer_mode,
         )
+        print(f"Uploaded dataset {uploaded_dataset.name} to {uploaded_dataset.url}")